requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
pydantic
python-dotenv
uuid
